Improve documentation for cacheDir and docker volume driver
There was no help function in jenkins for 'cache dirs' and I had to look into the code to understand how it worked.
The docker volume driver to where the documentation refers does not seem to be maintained anymore. Using the local docker volume driver is a valid alternative for a wide range of use cases .e.g. caching repositories