Add option to specify volumeDirs in the docker agent template
While cacheDirs is more focused on a unique cache per job, the volumeDirs provides a cache which is usable across multiple jobs. Giving us the opportunity to save disk space.
This is different from CacheDirs as we are able to specify the volume name, hence it is reusable between jobs and even between docker images.